+++ components/Tabs/TabPanel.tsx
@@ -1,0 +1,10 @@
+function TabPanel({ label, children }: any): JSX.Element {
+
+  return (
+    <div>
+      <span>{ label }</span>
+    </div>
+  );
+}
+
+export default TabPanel;