+++ containers/Product/ProductReview.tsx
@@ -1,0 +1,20 @@
+//components
+import { Tabs, Tab } from "../../components/Tabs";
+
+
+function ProductReview() {
+
+
+  return (
+    <div>
+      <Tabs>
+        <Tab label="Panel 1">Lorem ipsum dolor sit amet consectetur, adipisicing elit. Sint id fugiat praesentium culpa incidunt quae eligendi non corporis nisi, sequi minus natus maxime soluta neque perferendis rerum laboriosam assumenda repellat.</Tab>
+        <Tab label="Panel 2">Test</Tab>
+        <Tab label="Panel 3"/>
+      </Tabs>
+      ProductReview
+    </div>
+  );
+}
+
+export default ProductReview;