+++ containers/index.ts
@@ -1,1 +1,2 @@
-export { default as ProductList } from "./ProductsList";
+export { default as ProductList } from "./Product/ProductsList";
+export { default as ProductReview } from "./Product/ProductReview";